# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PenFlow API
  description: Core backend API for the PenFlow platform.
  version: 1.0.0
paths:
...
  /api/v1/engagements/{engagement_id}/report:
    get:
      tags:
      - Engagements
      summary: Get engagement report
      operationId: get_engagement_report_api_v1_engagements__engagement_id__report_get
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      - name: version
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Engagement Report Api V1 Engagements  Engagement
                  Id  Report Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/findings/{finding_id}:
    get:
      tags:
      - Findings
      summary: Get Finding
      operationId: get_finding_api_v1_findings__finding_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: finding_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Finding Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Findings
      summary: Update Finding
      operationId: update_finding_api_v1_findings__finding_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: finding_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Finding Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindingUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Findings
      summary: Delete Finding
      operationId: delete_finding_api_v1_findings__finding_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: finding_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Finding Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/findings/{finding_id}/evidence:
    post:
      tags:
      - Findings
      summary: Upload Finding Evidence
      operationId: upload_finding_evidence_api_v1_findings__finding_id__evidence_post
      security:
      - HTTPBearer: []
      parameters:
      - name: finding_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Finding Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_finding_evidence_api_v1_findings__finding_id__evidence_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceFileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/findings/{finding_id}/verify:
    patch:
      tags:
      - Findings
      summary: Verify Finding
      operationId: verify_finding_api_v1_findings__finding_id__verify_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: finding_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Finding Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reports/{report_id}/download:
    get:
      tags:
      - Reports
      summary: Download engagement report
      operationId: download_report_api_v1_reports__report_id__download_get
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_finding_evidence_api_v1_findings__finding_id__evidence_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_finding_evidence_api_v1_findings__finding_id__evidence_post
...
    EvidenceFileResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        file_name:
          type: string
          title: File Name
        mime_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Mime Type
        uploaded_at:
          type: string
          format: date-time
          title: Uploaded At
      type: object
      required:
      - id
      - file_name
      - uploaded_at
      title: EvidenceFileResponse
...
    FindingDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        engagement_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Engagement Id
        engagement_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Engagement Asset Id
        source:
          type: string
          title: Source
        status:
          $ref: '#/components/schemas/FindingStatus'
        is_verified:
          type: boolean
          title: Is Verified
        severity:
          $ref: '#/components/schemas/Severity'
        cvss_score:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Cvss Score
        cve_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cve Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
        asset_identifier:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset Identifier
        recommendation:
          anyOf:
          - type: string
          - type: 'null'
          title: Recommendation
        created_by:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Created By
        evidence_files:
          items:
            $ref: '#/components/schemas/EvidenceFileResponse'
          type: array
          title: Evidence Files
          default: []
      type: object
      required:
      - id
      - source
      - status
      - is_verified
      - severity
      - title
      - created_at
      title: FindingDetail
...
    FindingStatus:
      type: string
      enum:
      - open
      - in_progress
      - resolved
      - accepted_risk
      - false_positive
      title: FindingStatus
    FindingUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Title
        engagement_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Engagement Asset Id
        severity:
          anyOf:
          - $ref: '#/components/schemas/Severity'
          - type: 'null'
        cvss_score:
          anyOf:
          - type: number
            maximum: 10.0
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Cvss Score
        cve_id:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Cve Id
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        recommendation:
          anyOf:
          - type: string
          - type: 'null'
          title: Recommendation
        status:
          anyOf:
          - $ref: '#/components/schemas/FindingStatus'
          - type: 'null'
        is_verified:
          type: boolean
          title: Is Verified
      type: object
      required:
      - is_verified
      title: FindingUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Severity:
      type: string
      enum:
      - info
      - low
      - medium
      - high
      - critical
      title: Severity
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer